requirements.txt
Django~=5.2.1
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.3.1
drf-spectacular~=0.28.0
python-dotenv~=1.1.0